==> R/P0_CellMix_GAP_FEP.Rmd <==
---
title: "P0_CellMix_GAP_FEP"
author: "DJL"
date: "10 October 2017"
output: pdf_document
---

#CellMix approach:

##Install and load Libraries:
```{r Libraries, message=FALSE,warning=FALSE}


library(CellMix)
library(lumi)
library(reshape)
library(plyr)
library(dplyr)
library(ggplot2)


```



##Set directories:
```{r directory}
getwd()
data_dir <-"./data/"
P0_output_dir <-"./P0_Characterise/output/"
P0_figs_dir <-"./P0_Characterise/figs/"

```


##Load gene expression data (Should be LumiBatch object):
```{r }

lumidata<-"GAP_FEP_Full_Gene_Expression_Data_Linear.RData"
load(paste(data_dir,lumidata,sep=""))


```
